Move Cypress tests into descriptive folders
