+++ cypress/integration/individual-pages/aboutTheAircraft.spec.ts
@@ -10,7 +10,7 @@ import {
   whenIClickContinue,
   whenIClickOnTheErrorSummaryLinkContaining,
   whenIType,
-} from "./common.spec";
+} from "../common.spec";
 
 describe("As a beacon owner, I want to submit information about my aircraft", () => {
   const previousPageUrl = "/register-a-beacon/activity";
