+++ cypress/integration/individual-pages/aboutTheVessel.spec.ts
@@ -10,7 +10,7 @@ import {
   whenIClickContinue,
   whenIClickOnTheErrorSummaryLinkContaining,
   whenIType,
-} from "./common.spec";
+} from "../common.spec";
 
 describe("As a beacon owner, I want to submit information about my beacon", () => {
   const thisPageUrl = "/register-a-beacon/about-the-vessel";
