+++ cypress/integration/individual-pages/aircraftCommunications.spec.ts
@@ -9,7 +9,7 @@ import {
   thenTheUrlShouldContain,
   whenIClickOnTheErrorSummaryLinkContaining,
   whenIType,
-} from "./common.spec";
+} from "../common.spec";
 
 describe("As a beacon owner, I want to register details about the aircraft communications", () => {
   const pageUrl = "/register-a-beacon/aircraft-communications";
