+++ cypress/integration/individual-pages/moreDetails.spec.ts
@@ -8,7 +8,7 @@ import {
   thenTheUrlShouldContain,
   whenIClickOnTheErrorSummaryLinkContaining,
   whenIType,
-} from "./common.spec";
+} from "../common.spec";
 
 describe("As a beacon owner I want to submit more information about my beacon", () => {
   const thisPageUrl = "/register-a-beacon/more-details";
