+++ cypress/integration/individual-pages/purpose.spec.ts
@@ -1,4 +1,4 @@
-import { PageURLs } from "../../src/lib/urls";
+import { PageURLs } from "../../../src/lib/urls";
 import {
   andIClickContinue,
   andIHaveEnteredNoInformation,
@@ -7,7 +7,7 @@ import {
   requiredFieldErrorMessage,
   thenIShouldSeeAnErrorMessageThatContains,
   whenIClickContinue,
-} from "./common.spec";
+} from "../common.spec";
 
 describe("As a beacon owner, I want to submit the purpose for my beacon", () => {
   it("displays an error if no beacon use purpose is selected", () => {
